requirements.txt
aiogram==3.22.0
python-dotenv==1.1.1
aiohttp==3.12.15
aiofiles==24.1.0
pydantic==2.11.7
SQLAlchemy==2.0.36
psycopg2-binary==2.9.9
apscheduler==3.10.4
telethon==1.37.0
dateparser==1.2.0
requests==2.32.3
beautifulsoup4==4.12.3
lxml==5.3.0
easyocr==1.7.1
